10.4

+++ code/shapelet/src/supervised/Tools.java
@@ -6,7 +6,7 @@ import java.util.HashMap;
 import unsupervised.Metric;
 
 public class Tools {
-	public static double FStatistic(HashMap<Integer, ArrayList<Integer>> map){
+	public static double FStatistic(HashMap<Integer, ArrayList<Double>> map){
 		return Metric.FStatistic(map);
 	}
